Add support for modularized blueprints. Closes #10. Fixes #7.
This changes the way we use the blueprints, WITHOUT RETROCOMPATIBILITY.
Not there's an `index.js` file acting as a core for the package, and all
the blueprints and utils are exposed through this file.

var sailsEmberBlueprints = require('sails-ember-blueprints');
var blueprints = sailsEmberBlueprints.blueprints;
// {
//   create: function,
//   destroy: function,
//   find: funtion,
//   findone: function,
//   populate: function,
//   update: function
// }

var emberUtils = sailsEmberBlueprints.utils;

Check out the `README.md` file to know more about this changes.

Fix README.md markdown issues
