+++ blueprints/find.js
@@ -1,8 +1,8 @@
 /**
  * Module dependencies
  */
-var util = require( 'util' ),
-  actionUtil = require( './_util/actionUtil' );
+var util = require( 'util' );
+var emberUtils = require('./utils/emberUtils.js');
 
 /**
  * Enable sideloading. Edit config/blueprints.js and add:
@@ -36,7 +36,7 @@ var performSideload = (sails.config.blueprints.ember && sails.config.blueprints.
 module.exports = function findRecords( req, res ) {
 
   // Look up the model
-  var Model = actionUtil.parseModel( req );
+  var Model = emberUtils.parseModel( req );
 
   /* ENABLE if needed ( see https://github.com/mphasize/sails-ember-blueprints/issues/3 )
    * ----------------
@@ -45,18 +45,18 @@ module.exports = function findRecords( req, res ) {
    * of the `id` param.   (mainly here for compatibility for 0.9, where
    * there was no separate `findOne` action)
    */
-  // if ( actionUtil.parsePk( req ) ) {
+  // if ( emberUtils.parsePk( req ) ) {
   //  return require( './findone' )( req, res );
   // }
 
   // Lookup for records that match the specified criteria
   var query = Model.find()
-    .where( actionUtil.parseCriteria( req ) )
-    .limit( actionUtil.parseLimit( req ) )
-    .skip( actionUtil.parseSkip( req ) )
-    .sort( actionUtil.parseSort( req ) );
+    .where( emberUtils.parseCriteria( req ) )
+    .limit( emberUtils.parseLimit( req ) )
+    .skip( emberUtils.parseSkip( req ) )
+    .sort( emberUtils.parseSort( req ) );
 
-  query = actionUtil.populateEach( query, req );
+  query = emberUtils.populateEach( query, req );
   query.exec( function found( err, matchingRecords ) {
     if ( err ) return res.serverError( err );
 
@@ -69,10 +69,10 @@ module.exports = function findRecords( req, res ) {
       }
       // Also subscribe to instances of all associated models
       _.each( matchingRecords, function ( record ) {
-        actionUtil.subscribeDeep( req, record );
+        emberUtils.subscribeDeep( req, record );
       } );
     }
 
-    res.ok( actionUtil.emberizeJSON( Model, matchingRecords, req.options.associations, performSideload ) );
+    res.ok( emberUtils.emberizeJSON( Model, matchingRecords, req.options.associations, performSideload ) );
   } );
 };
