+++ package.json
@@ -2,7 +2,7 @@
   "name": "sails-ember-blueprints",
   "version": "0.1.4",
   "description": "Sails blueprints for a REST interface compatible with Ember Data's DS.RESTAdapter",
-  "main": "install.js",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -20,5 +20,9 @@
   "bugs": {
     "url": "https://github.com/mphasize/sails-ember-blueprints/issues"
   },
-  "homepage": "https://github.com/mphasize/sails-ember-blueprints"
+  "homepage": "https://github.com/mphasize/sails-ember-blueprints",
+  "dependencies": {
+    "lodash": "^2.4.1",
+    "pluralize": "^1.0.2"
+  }
 }
